MR-2096: Documents removed from edited submissions not downloadable on previous submission (#928)
* Pass plan package status to components with file uploading.

* Update deleteFile() to delete file if `DRAFT` submission or file not yet submitted.

* Update deleteFile() argument types.

* Add planPackageStatus prop and update arguments for deleteFile().

* Fix comment.

* Test for document delete button with mock

* function to list document keys and tests.

* Rollback changes to s3 functions.

* handle checking and deleting files from s3.

* Moved handling of file deletion to `StateSubmissionForm.tsx`

* Update test prop.

* Remove console log.

* Remove getKey.

* Add test for removing a file not in previous revisions.

* Move document test setup to apolloHelpers.tsx. Add test.

* `mockUnlockedHealthPlanPackageWithDocuments` function.

* Move file deletion function back into component.

* Move document helper functions to `documentHelper` directory.

* Move s3 delete file test to `StateSubmissionForm.test.ts`

* Change directory of `makeDataTable` import.

* Remove unnecessary checks and return Error.

* Move s3 file deletion function to child components.

* Fix tests

* Refactor `makeDocumentList` and add more tests.

* Fix `s3URL` in test.

+++ services/app-web/src/pages/StateSubmission/ContractDetails/ContractDetails.tsx
@@ -102,9 +102,11 @@ type FormError =
 export const ContractDetails = ({
     draftSubmission,
     showValidations = false,
+    previousDocuments,
     updateDraft,
 }: {
     draftSubmission: UnlockedHealthPlanFormDataType
+    previousDocuments: string[]
     showValidations?: boolean
     updateDraft: (
         input: UnlockedHealthPlanFormDataType
@@ -179,12 +181,20 @@ export const ContractDetails = ({
     }) => {
         setFileItems(fileItems)
     }
-    const handleDeleteFile = async (key: string) => {
-        const result = await deleteFile(key)
-        if (isS3Error(result)) {
-            throw new Error(`Error in S3 key: ${key}`)
-        }
 
+    const handleDeleteFile = async (key: string) => {
+        const isSubmittedFile =
+            previousDocuments &&
+            Boolean(
+                previousDocuments.some((previousKey) => previousKey === key)
+            )
+
+        if (!isSubmittedFile) {
+            const result = await deleteFile(key)
+            if (isS3Error(result)) {
+                throw new Error(`Error in S3 key: ${key}`)
+            }
+        }
         return
     }
